requirements.txt
Django==3.2.1
#keras==2.6.0
matplotlib==3.7.1
numpy==1.26.4
opencv_contrib_python==4.5.5.62
pandas==2.0.3
#pdf2image==1.16.0
Pillow==9.5.0
protobuf==3.20.3
scikit_image==0.23.2
scikit_learn==1.3.2
scipy==1.11.4
streamlit==0.81.1
tensorflow==2.16.1
tqdm==4.59.0
hachoir
# opencv_python==4.5.5.62